Handle nil properly when attempting to get type
There are nils with and without types (because Go). We were handling typed
nils, but an untyped nil gives an invalid reflect.Value. We need to check that
a Value is valid before calling `v.Type()` or other methods on it.

We also need to make sure that we compare a bare `nil` to a typed nil properly
in `d.Is` and `d.ValueIs`.